Cleaning up the code for status bar colors
